Create basic card modal component
