+++ src/components/game-board/card-modal/styles.module.scss
@@ -1,0 +1,19 @@
+@import '~styles/theme';
+
+.modal {
+  height: 100%;
+  width: 100%;
+}
+
+.overlay {
+  align-items: center;
+  background-color: rgba(200, 200, 200, 0.8);
+  bottom: 0;
+  display: flex;
+  justify-content: center;
+  left: 0;
+  position: fixed;
+  right: 0;
+  top: 0;
+  z-index: 1000;
+}
